backported pr#2065 Fix AcceptListGrpcQuerier concurrency issues
